+++ packages/core/database/migrations/2022_01_18_100001_add_customer_id_to_orders_table.php
@@ -1,0 +1,25 @@
+<?php
+
+use GetCandy\Base\Migration;
+use Illuminate\Database\Schema\Blueprint;
+use Illuminate\Support\Facades\Schema;
+
+class AddCustomerIdToOrdersTable extends Migration
+{
+    public function up()
+    {
+        Schema::table($this->prefix.'orders', function (Blueprint $table) {
+            $table->foreignId('customer_id')->after('id')
+                ->nullable()
+                ->constrained($this->prefix.'customers');
+        });
+    }
+
+    public function down()
+    {
+        Schema::table($this->prefix.'orders', function ($table) {
+            $table->dropForeign(['customer_id']);
+            $table->dropColumn('customer_id');
+        });
+    }
+}